intro-streamlit/requirements.txt
streamlit
Pillow
rembg